Make cli modules use only the public API
I had to move the save/writeProjectFile module to lib/projects to solve a circular dependency issue.
