+++ src/cli/commands/new.js
@@ -1,15 +1,12 @@
 const inquirer = require('inquirer')
-const chalk = require('chalk')
-const open = require('open')
-const writeProjectFile = require('../../save/writeProject')
-const projectTemplate = require('../../configs/template-project.json')
+const { createNewProject } = require('../../index')
 
 module.exports = {
   description: 'Create new project from scratch',
   optionalArgs: 'projectName',
   handler(projectName) {
     if (projectName) {
-      return newProject({ name: projectName })
+      return createNewProject({ name: projectName })
     }
 
     return askProjectName()
@@ -17,7 +14,7 @@ module.exports = {
 }
 
 function askProjectName() {
-  return getProjectDetails().then(newProject)
+  return getProjectDetails().then(createNewProject)
 }
 
 function getProjectDetails() {
@@ -36,14 +33,4 @@ function getProjectDetails() {
   ])
 }
 
-function newProject({ name, description }) {
-  return new Promise(function(resolve, reject) {
-    if (name) projectTemplate.project = name
-    if (description) projectTemplate.description = description
-    resolve(projectTemplate)
-  })
-    .then(writeProjectFile(name, description))
-    .then(filePath => {
-      open(filePath)
-    })
-}
+
