+++ src/index.js
@@ -1,10 +1,15 @@
 const open = require('./open')
 const save = require('./save')
-const { projectsDirPath, getAllProjects } = require('./lib/projects')
+const {
+  projectsDirPath,
+  getAllProjects,
+  createNewProject,
+} = require('./lib/projects')
 
 module.exports = {
   open,
   save,
   getAllProjects,
   projectsDirPath,
+  createNewProject,
 }
